res fixed a few warnings that occured on VC.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@26679 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/engine/renderloop/stdsteps/lightiter.cpp
@@ -296,7 +296,7 @@ void csLightIterRenderStep::Perform (iRenderView* rview, iSector* sector,
   //int nlights = lights->GetCount();
   const csArray<iLightSectorInfluence*>& lights = lightmgr->GetRelevantLights (sector,
       -1, false);
-  int nlights = lights.GetSize ();
+  size_t nlights = lights.GetSize ();
 
   csArray<iLight*> lightList (16);
